Refatoração e inicio dos desafios -Aula02
