+++ config.json
@@ -2,13 +2,15 @@
   "name": "Rafael Henrique (Donodo)",
   "job": "Desenvolvedor Front-end",
   "github": "rhpessoa",
-  "url_banner": "https://images.unsplash.com/photo-1516905041604-7935af78f572?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80",
+  "url_banner": "https://images.unsplash.com/photo-1487058792275-0ad4aaf24ca7?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80",
   "favoritos": {
     "AluraTubes Favoritos": [
       {
+        "id": "0",
         "user_nickname": "rhpessoa"
       },
       {
+        "id": "1",
         "user_nickname": "soulofpluto"
       }
     ]
